back to float from double

+++ Classes/reg.h
@@ -28,8 +28,8 @@
 typedef struct regTag reg;
 
 typedef struct {
-    double X;
-    double Y;
+    float X;
+    float Y;
 } regSample;
 
 #if defined(__cplusplus)
@@ -37,12 +37,12 @@ extern "C" {
 #endif
 
 extern reg*    regNew(regSample* sampleBuffer, size_t Samples);
-extern void    regPushSample(reg* pReg, double X, double Y);
+extern void    regPushSample(reg* pReg, float X, float Y);
 extern size_t  regGetSampleCount(const reg* pReg);
 extern void    regCalculateRegression(reg* pReg);
 extern void    regClearSamples(reg * pReg);
-extern double  regPredictY(const reg* pReg, double X);
-extern double  regSlope(const reg* pReg);
+extern float   regPredictY(const reg* pReg, float X);
+extern float   regSlope(const reg* pReg);
 
 #if defined(__cplusplus)
 }
